+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Upcoming
+* Food Consuming
+  * Added Stop consuming on hit. When the player is hit eating/drinking will reset
+
 ## 2.9.5
 * Added a priority to Hardness (should fix a problem with Tinkers Construct where you could mine blocks even if your mining speed was 0)
 * Fixed Forge not begin able to show the "missing dependency" screen due to Protection mixin (Thanks Mrbysco)